requirements.txt
flask
requests
transformers
torch
spacy
pyyaml
pytest